requirements.txt
audioread==3.0.1
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
colorama==0.4.6
comtypes==1.3.0
contourpy==1.2.0
coverage==5.5
cycler==0.12.1
decorator==5.1.1
deprecation==2.1.0
docutils==0.20.1
eyed3==0.9.7
Faker==23.2.1
filelock==3.13.1
filetype==1.2.0
fonttools==4.47.2
fsspec==2024.2.0
hmmlearn==0.3.0
idna==3.6
imbalanced-learn==0.12.0
imblearn==0.0
importlib-resources==6.1.1
Jinja2==3.1.3
joblib==1.3.2
keyboard==0.13.5
kiwisolver==1.4.5
lazy_loader==0.3
librosa==0.10.1
llvmlite==0.42.0
lockfile==0.12.2
MarkupSafe==2.1.5
matplotlib==3.8.2
mpmath==1.3.0
msgpack==1.0.7
mysql-connector==2.2.9
mysql-connector-python==8.0.33
networkx==3.2.1
numba==0.59.0
numpy==1.26.4
packaging==23.2
pillow==10.2.0
platformdirs==4.2.0
plotly==5.18.0
pooch==1.8.0
protobuf==3.20.3
PyAudio==0.2.14
pyAudioAnalysis==0.3.14
pycparser==2.21
pydub==0.25.1
pynndescent==0.5.11
pyparsing==3.1.1
pypiwin32==223
python-daemon==3.0.1
python-dateutil==2.8.2
python-dotenv==1.0.1
pyttsx3==2.90
pywin32==306
requests==2.31.0
Resemblyzer==0.1.4
scikit-learn==1.4.0
scipy==1.12.0
six==1.16.0
sounddevice==0.4.6
soundfile==0.12.1
soundmeter==0.1.5
soxr==0.3.7
SpeechRecognition==3.10.1
sympy==1.12
tenacity==8.2.3
threadpoolctl==3.2.0
toml==0.10.2
torch==2.2.0
tqdm==4.66.1
typing==3.7.4.3
typing_extensions==4.9.0
umap-learn==0.5.5
urllib3==2.2.0
v==1
webrtcvad==2.0.10
zipp==3.17.0
